requirement.txt
djangorestframework 
djangorestframework-simplejwt
django
gunicorn
whitenoise